+++ CHANGELOG.md
@@ -94,6 +94,7 @@
 - Conversion tool: updated mappings for RECAP output format.
 - Bitstamp parser: added "Staking reward" transaction type.
 - Crypto.com parser: added additional transaction kinds. ([#285](https://github.com/BittyTax/BittyTax/issues/285))
+- KuCoin parser: make direction field case-insensitive. ([#189](https://github.com/BittyTax/BittyTax/issues/189))
 
 ## Version [0.5.0] Beta (2021-11-11)
 Important:-